requirements.txt
streamlit>=1.30
supabase>=2.0
openai>=1.0
slack-sdk>=3.0
notion-client>=2.0
python-dotenv>=1.0
bcrypt>=4.0
